47_amazon_price_tracker/requirements.txt
requests==2.32.3
bs4==0.0.2
python-dotenv==1.0.1